Allow null-weights mappings
The map format now allows to prefix a target atom name with a !. The
origin atom will be mapped to that target one, but with a weight of 0.

This allow to ignore some atoms when setting the position, while still
using them for  graph purposes. It is especially needed for elnedyn
where only the CA is used to place the BB bead, but the rest of the
backbone is required to have an edge between consecutive BB beads.