Make PDB reader more flexible
Allow the resname to be 4 character by reading the empty column after
the resname as part of the resname. This does not affect well formated
PDB file, but it allows to read PDB files that make use of that empty
column such as those from the ATB.

This is a common variation around the PDB format, and it is supported at
least by pymol and VMD.

Note that the writing is not modified, so 4 chars resnames are still
truncated at writing to write valid PDB files.