Reduce the need to load all the force fields
We do not systematically load all the native force fields when we import
vermouth.forcefield. Instead we load them all only if we need one of
them. It is not an ideal solution, but it will reduce the test coverage
noise when running individual test cases. The noise is still there when
running the full test suite as some tests do request for a specific
native force field, therefore triggering the force field parsing
mechanism.